feat: implement schema generation for basic types and enums, enhance struct conversion logic
